implementations/relay/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-relay"
version = "0.1.0"
description = "Aether relay implementation"
requires-python = ">=3.10"
dependencies = [
  "PyNaCl",
  "blake3",
  "aioquic",
  "py-libp2p",
  "python-rocksdb",
  "websockets",
]

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "pytest",
  "PyYAML",
  "ruff",
  "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aether_relay*"]
